fix(stage6): align Alembic with ETLBatchUpload model, clean old migration, skip non-ORM services
